+++ Team6.[FBusSchedule].API/Team6.[FbusSchedule].Repository/EntityModel/DetailTrip.cs
@@ -5,9 +5,9 @@ namespace Team6._FbusSchedule_.Repository.EntityModel;
 
 public partial class DetailTrip
 {
-    public long TripId { get; set; }
+    public int TripId { get; set; }
 
-    public long StationId { get; set; }
+    public int StationId { get; set; }
 
     public DateTime? ArrivalTime { get; set; }
