+++ Team6.[FBusSchedule].API/Team6.[FbusSchedule].Repository/EntityModel/Station.cs
@@ -5,7 +5,7 @@ namespace Team6._FbusSchedule_.Repository.EntityModel;
 
 public partial class Station
 {
-    public long StationId { get; set; }
+    public int StationId { get; set; }
 
     public string StationName { get; set; }
